requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==2.0
django-background-tasks==1.1.13
django-celery==3.2.2
django-compat==1.0.15
django-dotenv==1.4.2
djangorestframework==3.8.2
kombu==3.0.37
psycopg2==2.7.4
pytz==2018.4
redis==2.10.6
six==1.11.0
